Added extra word list functionality
(startGame.cpp) Added a function to calculate the size of the word list instead of it being done manually(which may cause errors). Also made small adjustments to existing functions to incorporate new changes.

(Note: This change would make it easier for the user to add and delete words from the game as desired(via /words/word_directory.txt) to more easily configure the game.)